Added TC2.1, TC2.2, TC2.3 and TC3. Changed decimal to double for formatting.

+++ src/server/ConsoleAmountRetriever.cs
@@ -4,9 +4,9 @@ namespace server
 {
     class ConsoleAmountRetriever
     {
-        public virtual decimal GetAmount()
+        public virtual double GetAmount()
         {
-            return decimal.Parse(Console.ReadLine());
+            return double.Parse(Console.ReadLine());
         }
     }
 }
